Updated /index.php
Added a way to read a CSS file using a $_REQUEST. The script also checks
if there is a previously created cache file. If there is, it includes
it. If there is no cache file, then it uses the str_compress_css()
function to minimize the stylesheet and prints the $output.

At this point the project is totally functional but not ready for
release because I have no documentation for it.